requirements.txt
requests>=2.20.1
privex-jsonrpc>=1.1.2
privex-helpers>=1.3.3
golos-python>=1.0.3
python-dateutil
pytz
attrs

# Unit testing related
pytest
pytest-cov
coverage
codecov
Faker

# Privex/rpcemulator for emulating JsonRPC nodes
rpcemulator
